+++ packages/jest-circus/src/legacy-code-todo-rewrite/jestAdapterInit.ts
@@ -178,7 +178,7 @@ export const runAndTransformResultsToJestFormat = async ({
           : ancestorTitles.join(' '),
         invocations: testResult.invocations,
         location: testResult.location,
-        numPassingAsserts: 0,
+        numPassingAsserts: testResult.numPassingAsserts,
         retryReasons: testResult.retryReasons,
         status,
         title: testResult.testPath[testResult.testPath.length - 1],
@@ -238,6 +238,7 @@ const eventHandler = async (event: Circus.Event) => {
       break;
     }
     case 'test_done': {
+      event.test.numPassingAsserts = jestExpect.getState().numPassingAsserts;
       _addSuppressedErrors(event.test);
       _addExpectedAssertionErrors(event.test);
       break;